refactor(player): move stream mapping, selection and fetching out of EnhancedPlayerManager
